chore(refactor): move folders & component refactor. refs #30
