+++ src/App.tsx
@@ -1,11 +1,40 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
+import AddchartOutlinedIcon from '@mui/icons-material/AddchartOutlined';
+import BarChartOutlinedIcon from '@mui/icons-material/BarChartOutlined';
 import { Box } from '@mui/material'
 
-import DashboardTabs from './components/DashboardTabs/DashboardTabs'
+import Dashboard from './components/Dashboards/Dashboard';
+import DashboardPanel from './components/Dashboards/DashboardPanel';
+import Tabs from './components/Tabs/Tabs';
+import demoData from './config/widgets/demo.json';
 import { WidgetsProvider } from './context/WidgetsContext'
+import type { WidgetProps } from './types/types';
 
-const queryClient = new QueryClient()
+const TAB_CONFIG = [
+  {
+    label: 'Demo',
+    icon: <BarChartOutlinedIcon />,
+    content: 
+      <DashboardPanel title='Demo'>
+        <Dashboard
+          editable={false}
+          defaultData={demoData as unknown as WidgetProps[]}
+        />
+      </DashboardPanel>,
+  },
+  {
+    label: 'My Dashboard',
+    icon: <AddchartOutlinedIcon />,
+    content:
+      <DashboardPanel title='My Dashboard'>
+        <Dashboard editable={true}/>
+      </DashboardPanel>,
+  },
+];
+
+
+const queryClient = new QueryClient();
 
 function App() {
   return (
@@ -21,7 +50,7 @@ function App() {
         }}
       >
         <WidgetsProvider>
-          <DashboardTabs/>
+          <Tabs tabs={TAB_CONFIG} />
         </WidgetsProvider>
       </Box>
     </QueryClientProvider>
