+++ src/components/utils/validateWidgetJson.ts
@@ -1,5 +1,15 @@
-import type { WidgetProps } from '../common/types';
+import type { WidgetProps } from '../../types/types';
 
+/**
+ * Parses and validates a JSON structure representing one or more widgets
+ *
+ * This utility ensures that the provided data has the minimal structure
+ * required for widget configuration, including basic type checks for
+ * common widget properties
+ *
+ * @param jsonData - JSON string or object containing widget definitions
+ * @returns An array of validated widgets, or `null` if invalid
+ */
 export function validateWidgetJson(jsonData: unknown): WidgetProps[] | null {
   try {
     const parsed = typeof jsonData === 'string' ? JSON.parse(jsonData) : jsonData;
